+++ www/go/core/model/User.php
@@ -21,6 +21,8 @@ use go\core\jmap\Entity;
 use go\core\orm\CustomFieldsTrait;
 use go\core\util\DateTime;
 use go\core\validate\ErrorCode;
+use go\modules\community\addressbook\model\AddressBook;
+use go\modules\community\notes\model\NoteBook;
 
 
 class User extends Entity {
@@ -150,6 +152,12 @@ class User extends Entity {
 	
 	
 	public $max_rows_list;
+
+	/**
+	 *
+	 * @var bool
+	 */
+	protected $archive = false;
 	
 	/**
 	 * The user timezone
@@ -259,13 +267,18 @@ class User extends Entity {
 		$this->getPersonalGroup()->setValues($values);
 	}
 	
-	public function setValues(array $values) {
+	public function setValues(array $values)
+	{
 		$this->passwordVerified = false;
-		if(isset($values['archive'])) {
-			$this->archiveUser($values['archive']);
-			unset($values['archive']);
-		}
 		return parent::setValues($values);
+	}
+
+
+	public function setArchive($v) {
+		if(!go()->getAuthState()->isAdmin()) {
+			throw new Forbidden("Only admins can archive");
+		}
+		$this->archive = $v;
 	}
 
 	protected function canCreate()
@@ -649,6 +662,10 @@ class User extends Entity {
 		if($this->isNew()) {
 			$this->legacyOnSave();	
 		}
+
+		if($this->archive) {
+			$this->archiveUser();
+		}
 		
 		return true;		
 	}
@@ -932,23 +949,20 @@ class User extends Entity {
 	 * If a user is archived, any shares with themselves and non-admin users are deleted.Please note that we only do
 	 * this for community items. It is not entirely certain for other objects if they should be archived.
 	 *
-	 * @param bool $doArchive
 	 */
-	private function archiveUser(bool $doArchive = false)
+	private function archiveUser()
 	{
-		$userId = $this->id;
-
-		// Admins should not be able to be archived!
-		if (!$doArchive || $userId === go()->getUserId() || $this->isAdmin()) {
-			return;
-		}
 		$aclIds = [];
 
 		if ($defAddressBookId = $this->addressBookSettings->getDefaultAddressBookId()) {
-			$aclIds[] = \go\modules\community\addressbook\model\AddressBook::findById($defAddressBookId)->findAclId();
+			$addressBook = AddressBook::findById($defAddressBookId);
+			$aclIds[] = $addressBook->findAclId();
+			AddressBook::entityType()->change($addressBook);
 		};
 		if ($defNoteBookId = $this->notesSettings->getDefaultNoteBookId()) {
-			$aclIds[] = \go\modules\community\notes\model\NoteBook::findById($defNoteBookId)->findAclId();
+			$noteBook = NoteBook::findById($defNoteBookId);
+			$aclIds[] = $noteBook->findAclId();
+			NoteBook::entityType()->change($noteBook);
 		}
 		if ($defTaskListId = $this->taskSettings->default_tasklist_id) {
 			$aclIds[] = \GO\Tasks\Model\Tasklist::model()->findByPk($defTaskListId)->findAclId();