+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Core: Support --debug flag for cli.php
 - ldapauth: group member sync for ActiveDirectory broken
 - projects: fixed error when you didn't have access to the contacts linked to projects
+- Core: New feature to "Archive" users.
 
 27-08-2020 6.4.165
 - Time registration: fixed Firefox bug in Timesheet